refactor(reduce,planner): dedup agg-spec emission, share PK decode; clear clippy
planner: extract push_agg_specs as the single source of truth for the engine
agg_specs layout, shared by the SELECT projection and the HAVING-only aggregate
materialisation so the two stay in lockstep (notably the AVG → SUM+COUNT_NON_NULL
invariant, which reduce-output column positions depend on). Bundle bind_having_expr's
invariant parameters into a HavingCtx struct and drop the dead reduce_schema
pass-through (binding never read it; the caller compiles against it separately),
removing the too_many_arguments suppression.

wire: add decode_pk_column_owned (sibling of decode_pk_column) returning an owned
16-byte buffer, and call it from the three reduce-path sites that repeated the
scratch-buffer decode dance (reduce emit_reduce_row, emit_finalized_row, agg
step_from_batch).

clippy: clear the gnitz-engine deny-lint errors that had accumulated — mostly
stale `as usize` casts after batch offsets became usize, plus identity ops,
map_or→is_none_or, a needless range loop, items-after-test-module ordering, and
a complex-tuple type alias (ExchangeSideMeta). Load-bearing patterns kept and
annotated: Vec<Box<Table>> (raw-pointer stability across reallocation) and the
UnsafeCell-backed as_partitioned_mut (mut_from_ref). Also two gnitz-core
warnings (get(0)→first, manual checked_div).

+++ crates/gnitz-engine/src/catalog/store.rs
@@ -923,7 +923,7 @@ impl CatalogEngine {
         let end_pk = pack_column_id(owner_id + 1, 0);
         let mut cursor = self.sys_columns.open_cursor();
         // sys_columns has a single U64 PK; OPK == big-endian.
-        cursor.cursor.seek_bytes(&(start_pk as u64).to_be_bytes());
+        cursor.cursor.seek_bytes(&start_pk.to_be_bytes());
 
         let mut defs = Vec::new();
         while cursor.cursor.valid {
